requirements.txt
astroid==2.3.3
boto==2.49.0
boto3==1.10.28
botocore==1.13.28
certifi==2019.11.28
chardet==3.0.4
Cython==0.29.14
docopt==0.6.2
docutils==0.15.2
gensim==3.8.1
idna==2.8
isort==4.3.21
jmespath==0.9.4
joblib==0.14.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
nltk==3.4.5
numpy==1.17.4
pandas==0.25.3
praw==6.4.0
prawcore==1.0.1
pylint==2.4.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.3
six==1.13.0
smart-open==1.9.0
update-checker==0.16
urllib3==1.25.7
websocket-client==0.56.0
wrapt==1.11.2
